Move AppAgeText component to matters dir
